refactor : Member에 BaseTimeEntity 적용
#1
