+++ src/main/java/com/hyerijang/dailypay/member/MemberRepository.java
@@ -1,5 +1,6 @@
 package com.hyerijang.dailypay.member;
 
+import com.hyerijang.dailypay.member.entity.Member;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
