+++ src/main/java/com/hyerijang/dailypay/member/dto/RegisterRequest.java
@@ -1,6 +1,6 @@
 package com.hyerijang.dailypay.member.dto;
 
-import com.hyerijang.dailypay.member.Member;
+import com.hyerijang.dailypay.member.entity.Member;
 import jakarta.validation.constraints.NotBlank;
 import lombok.Getter;
 import lombok.Setter;
@@ -14,7 +14,7 @@ public class RegisterRequest {
 
     @NotBlank
     private String password;
-    
+
     public Member toEntity() {
         return Member.builder()
             .account(this.account)
