+++ src/main/java/com/hyerijang/dailypay/member/entity/Member.java
@@ -1,5 +1,6 @@
-package com.hyerijang.dailypay.member;
+package com.hyerijang.dailypay.member.entity;
 
+import com.hyerijang.dailypay.common.entity.BaseTimeEntity;
 import jakarta.persistence.Column;
 import jakarta.persistence.Entity;
 import jakarta.persistence.GeneratedValue;
@@ -14,7 +15,7 @@ import lombok.NoArgsConstructor;
 @Getter
 @Entity
 @NoArgsConstructor(access = AccessLevel.PROTECTED)
-public class Member {
+public class Member extends BaseTimeEntity {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
